+++ php/TimeInterval.php
