+++ php/tests/InstructorCollegeClassIntegrationTest.php
@@ -2,12 +2,13 @@
 
 /*
 	Georgy Marrero.
-	August 9, 2015.
+	August 10, 2015.
 */
 
 // INCLUDES
-require_once 'Instructor.php';
-require_once 'AvailSchedule.php';
+require_once '../Instructor.php';
+require_once '../AvailSchedule.php';
+require_once '../CollegeClass.php';
 
 // SOME SETUPS:
 	// InstructorAvailSchedule #1
@@ -34,8 +35,6 @@ require_once 'AvailSchedule.php';
 	$availableSchedule->setAvailableTimePeriod( DaysOfWeek::Saturday, TimePeriod::EarlyMorning );
 	$availableSchedule->setAvailableTimePeriod( DaysOfWeek::Saturday, TimePeriod::Evening );
 
-	$availableSchedule->setAvailableTimePeriod( DaysOfWeek::Sunday, TimePeriod::LateMorningAndAfternoon );
-
 	// InstructorAvailSchedule # 2
 	$anotherSchedule = new InstructorAvailSchedule();
 	$anotherSchedule->setAvailableTimePeriod( DaysOfWeek::Monday, TimePeriod::EarlyMorning );
@@ -52,45 +51,56 @@ require_once 'AvailSchedule.php';
 	$anotherSchedule->setAvailableTimePeriod( DaysOfWeek::Thursday, TimePeriod::Night );
 
 // TESTING BEGINS
-echo "<h1>TESTING INSTRUCTOR CLASS</h1><br><br>";
+echo "<h1>TESTING Instructor with CollegeClass CLASS</h1><br><br>";
 
 // A: Default Constructor
 echo '<h2>A: Default constructor</h2>';
 // 1. Default Constructor.
 $instructorA = new Instructor();
 
-// 2. Print (empty)
-echo "<h3>- Printing empty:</h3>";
-$instructorA->printObject();
-
-// 3. Now set values.
+// 2. Now set values.
 $instructorA->setName('Mrs. Anna Wall');
+
 $instructorA->setAvailSchedule($availableSchedule);
 
-// 4. Print object
-echo "<h3>- Printing with data:</h3>";
+$history = new CollegeClass();
+$history->setTitle('History 180');
+$history->setCredits(CollegeClassCredits::Three);
+$history->setWeeklyFrequency(CollegeClassWeeklyFrequency::TwicePerWeek);
+$history->setPeriod(CollegeClassSemestrialPeriod::FullSemester);
+$instructorA->addClass($history);
+
+// 3. Print object
 $instructorA->printObject();
 
 // B: Overloaded Constructor
 echo '<h2>B: Overloaded constructor</h2>';
 // 1. Overloaded Constructor.
-$instructorB = Instructor::create()->withName("Mr. John Doe")->withSchedule($availableSchedule);
+$instructorB = Instructor::create()->withName("Mr. John Doe")->withSchedule($anotherSchedule);
 
-// 2. Print object
-echo "<h3>- Printing with data:</h3>";
-$instructorB->printObject();
+// 2. Add classes.
+$math = new CollegeClass();
+$math->setTitle('Math 180');
+$math->setCredits(CollegeClassCredits::Four);
+$math->setWeeklyFrequency(CollegeClassWeeklyFrequency::MondayToThursday);
+$math->setPeriod(CollegeClassSemestrialPeriod::FullSemester);
+$instructorB->addClass($math);
 
-// 3. Now change values.
-$instructorB->setName('Mr. Johnny');
-$instructorB->setAvailSchedule($anotherSchedule);
+$anotherMath = new CollegeClass();
+$anotherMath->setTitle('Math 185');
+$anotherMath->setCredits(CollegeClassCredits::Four);
+$anotherMath->setWeeklyFrequency(CollegeClassWeeklyFrequency::ThricePerWeek);
+$anotherMath->setPeriod(CollegeClassSemestrialPeriod::FullSemester);
+$instructorB->addClass($anotherMath);
 
-// 4. Print object
-echo "<h3>- Printing with less data:</h3>";
-$instructorB->printObject();
+$yetAnotherMath = new CollegeClass();
+$yetAnotherMath->setTitle('Math 280');
+$yetAnotherMath->setCredits(CollegeClassCredits::Four);
+$yetAnotherMath->setWeeklyFrequency(CollegeClassWeeklyFrequency::TwicePerWeek);
+$yetAnotherMath->setPeriod(CollegeClassSemestrialPeriod::FullSemester);
+$instructorB->addClass($yetAnotherMath);
 
-// 5. Unset schedule (should result on error)
-echo "<h3>- Set schedule to 'null' -- expecting error:</h3>";
-$instructorB->setAvailSchedule(null);
+// 3. Print object
 $instructorB->printObject();
 
 ?>
